+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## [Unreleased]
 
-## [1.0.2] - 2026-08-11
-
-## [1.0.1] - 2026-08-11
+- Compat for Dream Dive Tweaks' Endless runs
 
 ## [1.0.0] - 2026-08-11