+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.1.0] - 2026-08-19
+
 - Add Retry Run button, also works in a successful run.
 - Minor compat for Dream Dive Tweaks' Endless runs.
 - Fix the fade to black after Moros' appearance hiding the game over screen.
@@ -13,7 +15,8 @@
 
 - First version of the mod!
 
-[unreleased]: https://github.com/adi1998/GameOverScreen/compare/1.0.2...HEAD
+[unreleased]: https://github.com/adi1998/GameOverScreen/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/adi1998/GameOverScreen/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/adi1998/GameOverScreen/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/adi1998/GameOverScreen/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/adi1998/GameOverScreen/compare/b1c882517b583831c17c6b90faaf4dd7a6d36d34...1.0.0